refactor: Pages 관련 메서드를 PagesInfo로 이동
