+++ src/main/java/kr/codesquad/cafe/post/PostService.java
@@ -7,14 +7,12 @@ import kr.codesquad.cafe.post.dto.PostForm;
 import kr.codesquad.cafe.post.dto.SimplePostForm;
 import kr.codesquad.cafe.post.exception.PostNotFoundException;
 import kr.codesquad.cafe.user.domain.User;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
+
+import static kr.codesquad.cafe.global.PagesInfo.*;
 
 @Service
 @Transactional(readOnly = true)
@@ -23,26 +21,10 @@ public class PostService {
 
     public static final int MAIN_PAGE_SIZE = 15;
     public static final int PROFILE_PAGE_SIZE = 6;
-    public static final String CREATED_DATE_TIME = "createdDateTime";
     private final PostRepository postRepository;
 
     public PostService(PostRepository postRepository) {
         this.postRepository = postRepository;
-    }
-
-    public static List<SimplePostForm> toSimplePostForm(List<Post> posts) {
-        return posts
-                .stream()
-                .map(SimplePostForm::from)
-                .collect(Collectors.toList());
-    }
-
-    private static Pageable getPageable(int currentPage, int pageSize) {
-        return PageRequest.of(currentPage, pageSize, Sort.by(Sort.Direction.DESC, CREATED_DATE_TIME)).previous();
-    }
-
-    private static int getPages(double allCount) {
-        return (int) Math.ceil(allCount / MAIN_PAGE_SIZE);
     }
 
     @Transactional
